mock data and wip graph
