+++ frontend/pages/index.js
@@ -1,118 +1,42 @@
 import Head from 'next/head'
-import Image from 'next/image'
-import { Inter } from 'next/font/google'
-import styles from '@/styles/Home.module.css'
 import { useEffect, useState } from 'react'
 import { API_HOST } from '../config'
+import { GraphExplorer } from '../components/GraphExplorer'
 
-const inter = Inter({ subsets: ['latin'] })
+const MapModeToggle = ({mapMode, setMapMode}) => {
+  return <label className="relative inline-flex items-center mr-5 cursor-pointer">
+    <input type="checkbox" value="" className="sr-only peer" checked={mapMode} onChange={() => setMapMode(x => !x)} />
+    <div className="w-11 h-6 bg-gray-200 rounded-full peer peer-focus:ring-red-300 dark:peer-focus:ring-red-800 dark:bg-gray-700 peer-checked:after:translate-x-full peer-checked:after:border-white after:content-[''] after:absolute after:top-0.5 after:left-[2px] after:bg-white after:border-gray-300 after:border after:rounded-full after:h-5 after:w-5 after:transition-all dark:border-gray-600 peer-checked:bg-red-600"></div>
+    <span className="ml-3 text-sm font-medium text-gray-600">Map mode</span>
+  </label>;
+}
+
+const ListExplorer = () => {
+  return <div>
+    Here goes your code Andrea and Cristina!
+    </div>
+}
 
 export default function Home() {
-  const [apiResult, setApiResult] = useState({});
+  const [pieces, setPieces] = useState({});
   useEffect(() => {
-    fetch(`${API_HOST}/test`).then(r => r.json()).then(setApiResult)
+    fetch(`${API_HOST}/pieces`).then(r => r.json()).then(r => setPieces(r['array']))
   }, [])
+  const [mapMode, setMapMode] = useState(false)
   return (
     <>
       <Head>
-        <title>Create Next App</title>
+        <title>Can I Play It?</title>
         <meta name="description" content="Generated by create next app" />
         <meta name="viewport" content="width=device-width, initial-scale=1" />
-        <link rel="icon" href="/favicon.ico" />
+        <link rel="icon" href="/favicon.png" />
       </Head>
-      <main className={styles.main}>
-        <div className={styles.description}>
-          <p className="text-red-400">
-            {JSON.stringify(apiResult)}
-          </p>
-          <div>
-            <a
-              href="https://vercel.com?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-              target="_blank"
-              rel="noopener noreferrer"
-            >
-              By{' '}
-              <Image
-                src="/vercel.svg"
-                alt="Vercel Logo"
-                className={styles.vercelLogo}
-                width={100}
-                height={24}
-                priority
-              />
-            </a>
-          </div>
-        </div>
-
-        <div className={styles.center}>
-          <Image
-            className={styles.logo}
-            src="/next.svg"
-            alt="Next.js Logo"
-            width={180}
-            height={37}
-            priority
-          />
-        </div>
-
-        <div className={styles.grid}>
-          <a
-            href="https://nextjs.org/docs?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-            className={styles.card}
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            <h2 className={inter.className}>
-              Docs <span>-&gt;</span>
-            </h2>
-            <p className={inter.className}>
-              Find in-depth information about Next.js features and&nbsp;API.
-            </p>
-          </a>
-
-          <a
-            href="https://nextjs.org/learn?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-            className={styles.card}
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            <h2 className={inter.className}>
-              Learn <span>-&gt;</span>
-            </h2>
-            <p className={inter.className}>
-              Learn about Next.js in an interactive course with&nbsp;quizzes!
-            </p>
-          </a>
-
-          <a
-            href="https://vercel.com/templates?framework=next.js&utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-            className={styles.card}
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            <h2 className={inter.className}>
-              Templates <span>-&gt;</span>
-            </h2>
-            <p className={inter.className}>
-              Discover and deploy boilerplate example Next.js&nbsp;projects.
-            </p>
-          </a>
-
-          <a
-            href="https://vercel.com/new?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-            className={styles.card}
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            <h2 className={inter.className}>
-              Deploy <span>-&gt;</span>
-            </h2>
-            <p className={inter.className}>
-              Instantly deploy your Next.js site to a shareable URL
-              with&nbsp;Vercel.
-            </p>
-          </a>
-        </div>
+      <main className="min-h-screen flex flex-col">
+        <MapModeToggle mapMode={mapMode} setMapMode={setMapMode} />
+        { !mapMode && <ListExplorer />}
+        { mapMode && <div className="flex justify-center content-center flex-1 bg-white">
+          <GraphExplorer pieces={pieces} /> 
+          </div>}
       </main>
     </>
   )
